refactor(web): split AgentStep into picker essentials + options
Extract AgentPickerEssentials (agent grid + structured-view choice) and
AgentOptions (workflow preset, sandbox/auto-approve toggles, advanced
launch knobs) from the monolithic AgentStep, plus a shared Toggle. The
single-screen wizard (#2210) will render the two halves separately so the
picker stays up top and the options fold behind More options; AgentOptions
takes collapsibleAdvanced so it can render flat there. AgentStep stays as a
thin wrapper rendering both with collapsibleAdvanced, keeping its isolated
unit tests intact.
